Refactor: Simplify main component layout with utility classes
Just a small clean up to keep the layout more straightforward. Removed the PageMainContent import and replaced it with USWDS utility classes. Also, fixed the layout to keep the footer pushed to the bottom (there was a typo with the utility class already in use). In the future, we might want to migrate the current PageMainContent component in veda-ui to USWDS and potentially reuse it here, but for now, it remains a styled component in veda-ui that's not in active use here.
Additionally, there was a bit of back and forth with the tabIndex, added it back for accessibility (enable skip to main).